+++ common/src/main/java/net/berserker_rpg/client/BerserkerClient.java
@@ -2,6 +2,7 @@ package net.berserker_rpg.client;
 
 import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
 import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererRegistry;
+import net.berserker_rpg.BerserkerClassMod;
 import net.berserker_rpg.client.armor.CustomArmorRenderer;
 import net.berserker_rpg.client.effect.RageParticles;
 import net.berserker_rpg.client.effect.RageRenderer;
@@ -11,6 +12,7 @@ import net.berserker_rpg.effect.BerserkerEffects;
 import net.berserker_rpg.item.armor.Armors;
 import net.berserker_rpg.spell.BerserkerSpells;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.particle.*;
 import net.spell_engine.api.effect.CustomModelStatusEffect;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
@@ -20,6 +22,8 @@ import net.spell_engine.client.gui.SpellTooltip;
 
 import java.util.List;
 import java.util.function.Supplier;
+
+import static net.berserker_rpg.compat.CompatLoadingCheck.armoryLoadCheck;
 
 public class BerserkerClient {
     public static void  init(){
@@ -37,12 +41,9 @@ public class BerserkerClient {
         registerArmorRenderer(Armors.wildlingArmorSet, CustomArmorRenderer::wildling_armor);
         registerArmorRenderer(Armors.northlingArmorSet, CustomArmorRenderer::northling_armor);
         registerArmorRenderer(Armors.netheriteNorthlingArmorSet, CustomArmorRenderer::netherite_northling_armor);
-        /*
-        if (FabricLoader.getInstance().isModLoaded("armory_rpgs") || BerserkerClassMod.tweaksConfig.value.ignore_items_required_mods) {
-            registerArmorRenderer(ArmoryCompat.warlordArmorSet.armorSet(), CustomArmorRenderer::warlord_armor);
+        if (armoryLoadCheck()) {
+            registerArmorRenderer(Armors.warlordArmorSet.armorSet(), CustomArmorRenderer::warlord_armor);
         }
-        */
-
         CustomParticleStatusEffect.register(BerserkerEffects.RAGE.effect, new RageParticles(1));
         CustomModelStatusEffect.register(BerserkerEffects.RAGE.effect, new RageRenderer());
     }