+++ common/src/main/java/net/berserker_rpg/compat/CompatLoadingCheck.java
@@ -1,0 +1,10 @@
+package net.berserker_rpg.compat;
+
+import net.fabricmc.loader.api.FabricLoader;
+
+public class CompatLoadingCheck {
+    public static boolean armoryLoadCheck(){
+        return FabricLoader.getInstance().isModLoaded("armory_rpgs") | FabricLoader.getInstance().isDevelopmentEnvironment();
+    }
+}
+